added a few things

+++ includes/philo.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   philo.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: Helene <Helene@student.42.fr>              +#+  +:+       +#+        */
+/*   By: hlesny <hlesny@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/13 19:19:03 by Helene            #+#    #+#             */
-/*   Updated: 2023/08/09 22:49:10 by Helene           ###   ########.fr       */
+/*   Updated: 2023/09/02 18:37:35 by hlesny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,11 +23,11 @@
 typedef struct  s_data
 {
 	int				philos_count;
-	/* int ou autre type de donnée ? */
-	int				time_to_die;
-	int				time_to_eat;
-	int				time_to_sleep;
-	int 			number_of_times_each_philosopher_must_eat;
+	/* unsigned long ou autre type de donnée ? */
+	unsigned long				time_to_die;
+	unsigned long				time_to_eat;
+	unsigned long				time_to_sleep;
+	unsigned long 			number_of_times_each_philosopher_must_eat;
 	
 	/* equal amount of forks and philosophers */
 	pthread_mutex_t	*forks;
